requirements.txt
asgiref==3.3.4
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
Django==3.1.7
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.11.2
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
google-api-core==1.26.3
google-auth==1.29.0
google-cloud==0.34.0
google-cloud-texttospeech==2.3.0
googleapis-common-protos==1.53.0
grpcio==1.37.0
gTTS==2.2.2
idna==2.10
packaging==20.9
proto-plus==1.18.1
protobuf==3.15.8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pygame==2.0.1
pyglet==1.5.16
pyparsing==2.4.7
python-dotenv==0.17.0
pytz==2021.1
PyYAML==5.4.1
requests==2.25.1
rsa==4.7.2
six==1.15.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.26.4
